requirements.txt
ipykernel
openai==0.28
gradio
requests
python-dotenv